short circuit instead of throwing error, remove cast and move to service level to keep data fetcher clean
